git card money owed to store

+++ app/graphql/types/gift_card_change_type.rb
@@ -3,7 +3,7 @@ module Types
     field :id, ID, null: true
     field :store, StoreType, null: true
     field :gift_card, GiftCardType, null: true
-    field :change_value, Integer, null: true
+    field :change_value, Float, null: true
     field :change_event, String, null: true
   end
 end
